+++ DPEMoveDAL/ViewModels/GroupEditViewModel.cs
@@ -8,7 +8,7 @@ namespace DPEMoveDAL.ViewModels
 {
     public class GroupEditViewModel
     {
-        public int GroupId { get; set; }
+        public decimal GroupId { get; set; }
         public string GroupName { get; set; }
         public string Description { get; set; }
         public int Status { get; set; }
